Add the Time Series example.
From : https://apple.github.io/foundationdb/time-series.html

This was a good exercise which helped me see the value of
directories (vs nested subspaces) for maintaining related data.

Also, I need to add the mutation wrappers to make this example clean,
and I need to rethink the encoding / decoding API for keys and values
to make the functions easier to use.